Refactor TransferView Modal
